+++ Tests/NyrisSDKTests/MatchingServiceTests.swift
@@ -28,9 +28,9 @@ class MatchingServiceTests: XCTestCase {
         let expectations = expectation(description: "invalid size: size too small")
         let service = ImageMatchingService()
         
-        service.match(image: image) { (offers, error) in
+        service.match(image: image) { (product, error) in
             XCTAssertNotNil(error)
-            XCTAssertNil(offers)
+            XCTAssertNil(product)
             
             switch error! {
             case ImageError.invalidSize(let message):
@@ -54,9 +54,9 @@ class MatchingServiceTests: XCTestCase {
         let service = ImageMatchingService()
         // any language is fine here
         service.acceptLanguage = "*"
-        service.match(image: image) { (offers, error) in
-            XCTAssertNotNil(offers)
-            XCTAssertFalse(offers!.isEmpty)
+        service.match(image: image) { (offersResult, error) in
+            XCTAssertNotNil(offersResult)
+            XCTAssertFalse(offersResult!.products.isEmpty)
             XCTAssertNil(error)
             expectations.fulfill()
         }
@@ -66,7 +66,7 @@ class MatchingServiceTests: XCTestCase {
     // XOPTION tests
     
     func test_it_can_use_default_search_mode_with_results() {
-        guard let image = UIImage(named: "test-image.jpg", in: TestsHelper.bundle, compatibleWith: nil) else {
+        guard let image = UIImage(named: "product_test_512.png", in: TestsHelper.bundle, compatibleWith: nil) else {
             fatalError("not found")
         }
         
@@ -75,10 +75,10 @@ class MatchingServiceTests: XCTestCase {
         let limit = 20
         service.xOptions = "default"
         
-        service.match(image: image) { (offers, error) in
-            XCTAssertNotNil(offers)
+        service.match(image: image) { (offersResult, error) in
+            XCTAssertNotNil(offersResult)
             XCTAssertNil(error)
-            XCTAssertLessThanOrEqual(offers!.count, limit)
+            XCTAssertLessThanOrEqual(offersResult!.products.count, limit)
             
             expectations.fulfill()
             
@@ -87,7 +87,7 @@ class MatchingServiceTests: XCTestCase {
     }
     
     func test_it_can_limit_offers_to_10() {
-        guard let image = UIImage(named: "test-image.jpg", in: TestsHelper.bundle, compatibleWith: nil) else {
+        guard let image = UIImage(named: "product_test_512.png", in: TestsHelper.bundle, compatibleWith: nil) else {
             fatalError("not found")
         }
 
@@ -96,10 +96,10 @@ class MatchingServiceTests: XCTestCase {
         let limit = 10
         service.xOptions = "exact +similarity +ocr limit=\(limit)"
         
-        service.match(image: image) { (offers, error) in
-            XCTAssertNotNil(offers)
+        service.match(image: image) { (offersResult, error) in
+            XCTAssertNotNil(offersResult)
             XCTAssertNil(error)
-            XCTAssertLessThanOrEqual(offers!.count, limit)
+            XCTAssertLessThanOrEqual(offersResult!.products.count, limit)
             
             expectations.fulfill()
             
@@ -108,7 +108,7 @@ class MatchingServiceTests: XCTestCase {
     }
     
     func test_it_can_use_exact_search_mode_with_no_offers() {
-        guard let image = UIImage(named: "test-image.jpg", in: TestsHelper.bundle, compatibleWith: nil) else {
+        guard let image = UIImage(named: "product_test_512.png", in: TestsHelper.bundle, compatibleWith: nil) else {
             fatalError("not found")
         }
         
@@ -116,10 +116,10 @@ class MatchingServiceTests: XCTestCase {
         let service = ImageMatchingService()
         service.xOptions = "exact"
         
-        service.match(image: image) { (offers, error) in
-            XCTAssertNotNil(offers)
+        service.match(image: image) { (offersResult, error) in
+            XCTAssertNotNil(offersResult)
             XCTAssertNil(error)
-            XCTAssertEqual(offers?.count, 0)
+            XCTAssertEqual(offersResult!.products.count, 0)
             
             expectations.fulfill()
             
@@ -128,7 +128,7 @@ class MatchingServiceTests: XCTestCase {
     }
     
     func test_it_can_use_similarity_search_mode_limited_3_offers() {
-        guard let image = UIImage(named: "test-image.jpg", in: TestsHelper.bundle, compatibleWith: nil) else {
+        guard let image = UIImage(named: "product_test_512.png", in: TestsHelper.bundle, compatibleWith: nil) else {
             fatalError("not found")
         }
         
@@ -137,10 +137,10 @@ class MatchingServiceTests: XCTestCase {
         let limit = 3
         service.xOptions = "similarity limit=\(limit)"
         
-        service.match(image: image) { (offers, error) in
-            XCTAssertNotNil(offers)
+        service.match(image: image) { (offersResult, error) in
+            XCTAssertNotNil(offersResult)
             XCTAssertNil(error)
-            XCTAssertEqual(offers?.count, limit)
+            XCTAssertEqual(offersResult!.products.count, limit)
             
             expectations.fulfill()
             
@@ -150,7 +150,7 @@ class MatchingServiceTests: XCTestCase {
     
     
     func test_it_can_use_similarity_search_mode_limited_2_offers() {
-        guard let image = UIImage(named: "test-image.jpg", in: TestsHelper.bundle, compatibleWith: nil) else {
+        guard let image = UIImage(named: "product_test_512.png", in: TestsHelper.bundle, compatibleWith: nil) else {
             fatalError("not found")
         }
         
@@ -159,10 +159,10 @@ class MatchingServiceTests: XCTestCase {
         let limit = 2
         service.xOptions = "similarity limit=\(limit)"
         
-        service.match(image: image) { (offers, error) in
-            XCTAssertNotNil(offers)
+        service.match(image: image) { (offersResult, error) in
+            XCTAssertNotNil(offersResult)
             XCTAssertNil(error)
-            XCTAssertEqual(offers?.count, limit)
+            XCTAssertEqual(offersResult!.products.count, limit)
             
             expectations.fulfill()
             
@@ -172,7 +172,7 @@ class MatchingServiceTests: XCTestCase {
     
     
     func test_if_similarityThreshold_090_returns_1_offers() {
-        guard let image = UIImage(named: "test-image.jpg", in: TestsHelper.bundle, compatibleWith: nil) else {
+        guard let image = UIImage(named: "product_test_512.png", in: TestsHelper.bundle, compatibleWith: nil) else {
             fatalError("not found")
         }
         
@@ -183,10 +183,10 @@ class MatchingServiceTests: XCTestCase {
         
         service.xOptions = "similarity similarity.threshold=\(similarityThreshold)"
         
-        service.match(image: image) { (offers, error) in
-            XCTAssertNotNil(offers)
+        service.match(image: image) { (offersResult, error) in
+            XCTAssertNotNil(offersResult)
             XCTAssertNil(error)
-            XCTAssertEqual(offers?.count, limit)
+            XCTAssertEqual(offersResult!.products.count, limit)
             
             expectations.fulfill()
             
@@ -195,7 +195,7 @@ class MatchingServiceTests: XCTestCase {
     }
     
     func test_it_can_use_ocr_search_mode_with_no_offers() {
-        guard let image = UIImage(named: "test-image.jpg", in: TestsHelper.bundle, compatibleWith: nil) else {
+        guard let image = UIImage(named: "product_test_512.png", in: TestsHelper.bundle, compatibleWith: nil) else {
             fatalError("not found")
         }
         
@@ -203,14 +203,24 @@ class MatchingServiceTests: XCTestCase {
         let service = ImageMatchingService()
         service.xOptions = "ocr"
         
-        service.match(image: image) { (offers, error) in
-            XCTAssertNotNil(offers)
+        service.match(image: image) { (offersResult, error) in
+            XCTAssertNotNil(offersResult)
             XCTAssertNil(error)
-            XCTAssertEqual(offers?.count, 0)
+            XCTAssertEqual(offersResult!.products.count, 0)
             
             expectations.fulfill()
             
         }
         wait(for: [expectations], timeout: 40)
+    }
+    
+    func test_it_can_add_session_id_to_query() {
+        guard let image = UIImage(named: "product_test_512.png", in: TestsHelper.bundle, compatibleWith: nil) else {
+            fatalError("not found")
+        }
+        
+        let expectations = expectation(description: "Expected no items")
+        let service = ImageMatchingService()
+        
     }
 }